removed uptime command and added it to stats
